bugfix(core) Fixed serializer exceptions placement and added annotations
