+++ module/core/src/Application/Serializer/NormalizerInterface.php
@@ -9,9 +9,18 @@ declare(strict_types=1);
 
 namespace Ergonode\Core\Application\Serializer;
 
+use Ergonode\Core\Application\Exception\DenoralizationException;
+use Ergonode\Core\Application\Exception\NormalizationException;
+
 interface NormalizerInterface
 {
+    /**
+     * @throws NormalizationException
+     */
     public function normalize(object $data, ?string $type): array;
 
+    /**
+     * @throws DenoralizationException
+     */
     public function denormalize(array $data, string $type): object;
 }
